discord_openai_bot/requirements.txt
discord.py
openai
requests
python-dotenv
